+++ phenol-core/src/main/java/org/monarchinitiative/phenol/base/PhenolRuntimeException.java
@@ -5,7 +5,7 @@ package org.monarchinitiative.phenol.base;
  *
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
-public class OntoLibRuntimeException extends RuntimeException {
+public class PhenolRuntimeException extends RuntimeException {
 
   private static final long serialVersionUID = 1L;
 
@@ -15,7 +15,7 @@ public class OntoLibRuntimeException extends RuntimeException {
    * @param message Message of exception.
    * @param cause Causing {@link Throwable}.
    */
-  public OntoLibRuntimeException(String message, Throwable cause) {
+  public PhenolRuntimeException(String message, Throwable cause) {
     super(message, cause);
   }
 
@@ -24,7 +24,7 @@ public class OntoLibRuntimeException extends RuntimeException {
    *
    * @param message Message of exception.
    */
-  public OntoLibRuntimeException(String message) {
+  public PhenolRuntimeException(String message) {
     super(message);
   }
 }