+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ v0.1.7
 ------
 - refactored phenol to use JGraphT library
 - Adding OWLAPI based parser
+- Refactoring HPO Disease annotation parser
 
 ------
 v0.1.6